UHF-6203: Moved functional tests to correct namespace
